requirements.txt
"fastapi[standard]"
deep-translator
langchain-core
langchain
gtts
openai-whisper
pymupdf
sentence-transformers
chromadb

